fix(deps): Add python-venv dependency for pipx on Arch
The script was failing on Arch Linux during the 'pipx'
installation in '01-system-setup.sh'.

This was a symmetrical bug to one previously fixed on
Debian (related to 'ensurepip'). 'pipx' requires Python's
'venv' module to function. We had added 'python3-venv'
to the Debian list, but neglected to add the Arch
equivalent ('python-venv') to the 'pkgs_essentials_arch'
list.

This commit adds 'python-venv' to 'pkgs_essentials_arch'
to satisfy this dependency for 'pipx'.